Add some logging + refactor if else

+++ tests/test_repository.py
@@ -30,9 +30,13 @@ from huggingface_hub.repository import (
     is_tracked_upstream,
     is_tracked_with_lfs,
 )
+from huggingface_hub.utils import logging
 
 from .testing_constants import ENDPOINT_STAGING, PASS, USER
 from .testing_utils import set_write_permission_and_retry, with_production_testing
+
+
+logger = logging.get_logger(__name__)
 
 
 def repo_name(id=uuid.uuid4().hex[:6]):
@@ -64,10 +68,11 @@ class RepositoryTest(RepositoryCommonTest):
         cls._token = cls._api.login(username=USER, password=PASS)
 
     def setUp(self):
-        try:
-            shutil.rmtree(WORKING_REPO_DIR, onerror=set_write_permission_and_retry)
-        except FileNotFoundError:
-            pass
+        if os.path.exists(WORKING_REPO_DIR) and os.path.isdir(WORKING_REPO_DIR):
+            try:
+                shutil.rmtree(WORKING_REPO_DIR, onerror=set_write_permission_and_retry)
+            except FileNotFoundError:
+                pass
 
         self.REPO_NAME = repo_name()
         self._repo_url = self._api.create_repo(name=self.REPO_NAME, token=self._token)
@@ -97,6 +102,9 @@ class RepositoryTest(RepositoryCommonTest):
             pass
 
     def test_init_from_existing_local_clone(self):
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         subprocess.run(
             ["git", "clone", self._repo_url, WORKING_REPO_DIR],
             check=True,
@@ -113,6 +121,9 @@ class RepositoryTest(RepositoryCommonTest):
                 _ = Repository(tmpdirname)
 
     def test_init_clone_in_empty_folder(self):
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         repo = Repository(WORKING_REPO_DIR, clone_from=self._repo_url)
         repo.lfs_track(["*.pdf"])
         repo.lfs_enable_largefiles()
@@ -121,6 +132,9 @@ class RepositoryTest(RepositoryCommonTest):
         self.assertIn("random_file.txt", os.listdir(WORKING_REPO_DIR))
 
     def test_git_lfs_filename(self):
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         os.mkdir(WORKING_REPO_DIR)
         subprocess.run(
             ["git", "init"],
@@ -147,6 +161,9 @@ class RepositoryTest(RepositoryCommonTest):
     def test_init_clone_in_nonempty_folder(self):
         # Create dummy files
         # one is lfs-tracked, the other is not.
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         os.makedirs(WORKING_REPO_DIR, exist_ok=True)
         with open(os.path.join(WORKING_REPO_DIR, "dummy.txt"), "w") as f:
             f.write("hello")
@@ -158,6 +175,9 @@ class RepositoryTest(RepositoryCommonTest):
 
     def test_init_clone_in_nonempty_non_linked_git_repo(self):
         # Create a new repository on the HF Hub
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         temp_repo_url = self._api.create_repo(
             name=f"{self.REPO_NAME}-temp", token=self._token
         )
@@ -180,6 +200,9 @@ class RepositoryTest(RepositoryCommonTest):
         self._api.delete_repo(name=f"{self.REPO_NAME}-temp", token=self._token)
 
     def test_init_clone_in_nonempty_linked_git_repo_with_token(self):
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         Repository(
             WORKING_REPO_DIR, clone_from=self._repo_url, use_auth_token=self._token
         )
@@ -189,6 +212,9 @@ class RepositoryTest(RepositoryCommonTest):
 
     def test_init_clone_in_nonempty_linked_git_repo(self):
         # Clone the repository to disk
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         Repository(WORKING_REPO_DIR, clone_from=self._repo_url)
 
         # Add to the remote repository without doing anything to the local repository.
@@ -205,6 +231,9 @@ class RepositoryTest(RepositoryCommonTest):
 
     def test_init_clone_in_nonempty_linked_git_repo_unrelated_histories(self):
         # Clone the repository to disk
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         repo = Repository(
             WORKING_REPO_DIR,
             clone_from=self._repo_url,
@@ -230,6 +259,9 @@ class RepositoryTest(RepositoryCommonTest):
         Repository(WORKING_REPO_DIR, clone_from=self._repo_url)
 
     def test_add_commit_push(self):
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         repo = Repository(
             WORKING_REPO_DIR,
             clone_from=self._repo_url,
@@ -256,6 +288,9 @@ class RepositoryTest(RepositoryCommonTest):
         shutil.rmtree(WORKING_REPO_DIR)
 
     def test_add_commit_push_non_blocking(self):
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         repo = Repository(
             WORKING_REPO_DIR,
             clone_from=self._repo_url,
@@ -292,6 +327,9 @@ class RepositoryTest(RepositoryCommonTest):
         shutil.rmtree(WORKING_REPO_DIR)
 
     def test_context_manager_non_blocking(self):
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
         repo = Repository(
             WORKING_REPO_DIR,
             clone_from=self._repo_url,
